+++ lib/list_utilities.dart
@@ -1,5 +1,6 @@
 /// A collection of extension methods for [Iterable]s, [List]s, and [Set]s.
 library list_utilities;
 
+export 'src/iterable_utilities.dart';
 export 'src/list_utilities.dart';
 export 'src/set_utilities.dart';