+++ lib/src/set_utilities.dart
@@ -1,22 +1,13 @@
-/// /// Extends [Set] with additional basic methods.
+import 'iterable_utilities.dart';
+
+/// Extends [Set] with additional basic methods.
 extension SetUtilities<T> on Set<T> {
-  /// Returns `true` if [other] is the same length as `this` and
-  /// contains all of the same elements, regardless of their order.
-  bool equals(Iterable<T> other) {
-    assert(other != null);
-
-    if (length != other.length) {
-      return false;
-    }
-
-    final elements = toSet();
-    for (var element in other) {
-      if (!elements.remove(element)) {
-        return false;
-      }
-    }
-
-    return true;
+  /// Removes and returns a random value from the set.
+  T removeRandom([int seed]) {
+    assert(isNotEmpty, 'The list must not be empty.');
+    final element = random(seed);
+    remove(element);
+    return element;
   }
 
   /// Returns a new [Set] from `this` with [other]'s elements