+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## [0.1.1] - February 1, 2021
+
+* Moved [equals] from [SetUtilities] to [IterableUtilities], renamed
+it to [matches], and added the [ordered] parameter to it.
+
+* Moved [random] from [ListUtilities] to [IterableUtilities].
+
+* Added [removeRandom] to [ListUtilities] and [SetUtilities].
+
+* Added the [removeFirst], [removeFirstWhere] and [removeLastWhere]
+methods to [ListUtilities].
+
 ## [0.1.0] - January 28, 2021
 
 * Initial release.